requirements_rag.txt
# RAG System Dependencies
# Core LangChain components
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.28
langchain-core==0.3.75
langchainhub==0.1.21

# Text processing and tokenization
tiktoken==0.11.0

# Vector database
chromadb==1.0.20

# PDF processing
pypdf==4.2.0
pymupdf == 1.26.4

# HTTP requests for internet search
aiohttp==3.12.15
requests==2.32.5

# Data processing
numpy==2.2.6
bs4==0.0.2

# Async support
asyncio-throttle==1.0.2
spacy==3.8.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
docling
docling-core
yt_dlp
webvtt-py
translators
moviepy
langchain_chroma
langgraph
tavily-python
ffmpeg